working on

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: nhuang <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/20 10:22:09 by nhuang            #+#    #+#             */
-/*   Updated: 2023/11/27 19:40:22 by nhuang           ###   ########.fr       */
+/*   Updated: 2024/06/28 20:06:15 by nhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 #  define BUFFER_SIZE 1024
 # endif
 
-int		ft_strlen(const char *str);
+size_t		ft_strlen(const char *str);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
 char	*ft_strchr(const char *s, int i);
 char	*ft_strdup(const char *s);
